* transcode_data.h: return output functions ssize_t.
* transcode.c (transcode_restartable0): don't need to cast the result
  of output functions.

* enc/trans/newline.trans: follow the type change.

* enc/trans/escape.trans: ditto.

* enc/trans/utf_16_32.trans: ditto.

* enc/trans/iso2022.trans: ditto.

* enc/trans/japanese.trans: ditto.



git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@19351 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ transcode_data.h
@@ -84,9 +84,9 @@ struct rb_transcoder {
     int (*state_fini_func)(void*); /* ret==0:success ret!=0:failure(errno) */
     VALUE (*func_ii)(void*, VALUE); /* info  -> info   */
     VALUE (*func_si)(void*, const unsigned char*, size_t); /* start -> info   */
-    int (*func_io)(void*, VALUE, const unsigned char*, size_t); /* info  -> output */
-    int (*func_so)(void*, const unsigned char*, size_t, unsigned char*, size_t); /* start -> output */
-    int (*finish_func)(void*, unsigned char*, size_t); /* -> output */
+    ssize_t (*func_io)(void*, VALUE, const unsigned char*, size_t); /* info  -> output */
+    ssize_t (*func_so)(void*, const unsigned char*, size_t, unsigned char*, size_t); /* start -> output */
+    ssize_t (*finish_func)(void*, unsigned char*, size_t); /* -> output */
     int (*resetsize_func)(void*); /* -> len */
     int (*resetstate_func)(void*, unsigned char*, size_t); /* -> output */
 };
